Update PolyMc to 7b38391
