+++ build.gradle
@@ -41,6 +41,7 @@ allprojects {
         maven { url = "https://api.modrinth.com/maven" }
         maven { url 'https://oss.sonatype.org/content/repositories/snapshots' }
         maven { url = "https://maven.enginehub.org/repo/" }
+        maven { url = "https://maven.theepicblock.nl"; content { includeGroup("nl.theepicblock") } }
 
         mavenLocal()
     }
