+++ polymer-blocks/src/main/java/eu/pb4/polymer/ext/blocks/mixin/polymc/PolyRegistryMixin.java
@@ -1,9 +1,10 @@
 package eu.pb4.polymer.ext.blocks.mixin.polymc;
 
 import eu.pb4.polymer.ext.blocks.impl.BlockExtBlockMapper;
-import eu.pb4.polymer.mixin.compat.polymc_BlockStateManagerAccessor;
 import io.github.theepicblock.polymc.api.PolyRegistry;
+import io.github.theepicblock.polymc.api.SharedValuesKey;
 import io.github.theepicblock.polymc.api.block.BlockStateManager;
+import net.minecraft.block.Block;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Pseudo;
 import org.spongepowered.asm.mixin.Shadow;
@@ -14,13 +15,16 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Pseudo
 @Mixin(PolyRegistry.class)
 public abstract class PolyRegistryMixin {
-    @Shadow public abstract BlockStateManager getBlockStateManager();
+    @Shadow public abstract <T> T getSharedValues(SharedValuesKey<T> key);
 
     @Inject(method = "<init>", at = @At("TAIL"))
     private void polymer_block_catchRegistry(CallbackInfo ci) {
+        var blockStateManager = this.getSharedValues(BlockStateManager.KEY);
         for (var takenState : BlockExtBlockMapper.INSTANCE.stateMap.keySet()) {
             try {
-                ((polymc_BlockStateManagerAccessor) this.getBlockStateManager()).callRequestBlockState(takenState.getBlock(), state -> state == takenState, (x, y) -> { });
+                // Request this specific state from PolyMc
+                // This will cause PolyMc to mark it as used
+                blockStateManager.requestBlockState(state -> state == takenState, new Block[]{takenState.getBlock()}, (block, registry) -> { });
             } catch (Exception e) {
                 e.printStackTrace();
             }
